Update 014 - Sum by Factors.py
added another solution where instead of checking if a value is prime, we check if the value exists in a set containing prime numbers up to the max value in the lst, which is 1.55 faster than first implementation
NOTE: usage of max(lst) is as fast as (numpy.array(lst)).max() (0.045% difference). It seems that speed advantage gained by using numpy is lost while converting list to np.array.